Fix: 오류 수정 및 리팩터링

+++ src/main/java/lotto/applications/GameService.java
@@ -70,7 +70,7 @@ public class GameService {
         return Calculator.calculateProfit(money, getTotalAmountOfHit(prizeMap));
     }
 
-    private int getTotalAmountOfHit(Map<Place, Integer> prizeMap) {
+    private long getTotalAmountOfHit(Map<Place, Integer> prizeMap) {
         return Stream.of(prizeMap)
                 .flatMap(prize -> prize.entrySet()
                         .stream()
